requirements.txt
blinker==1.9.0
boolean.py==5.0
CacheControl==0.14.4
certifi==2024.7.4
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
cyclonedx-python-lib==11.7.0
defusedxml==0.7.1
filelock==3.25.2
Flask==3.1.3
Flask-SQLAlchemy==3.0.5
Flask-Limiter==4.1.1
Flask-WTF==1.2.2
greenlet==2.0.2
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.6
license-expression==30.4.4
mailchimp-marketing==3.0.80
markdown-it-py==4.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
msgpack==1.1.2
packageurl-python==0.17.6
packaging==26.0
pip-api==0.0.34
pip-requirements-parser==32.0.1
pip_audit==2.10.0
platformdirs==4.9.4
py-serializable==2.1.0
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.32.4
rich==14.3.3
six==1.16.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.20
tomli==2.4.0
tomli_w==1.2.0
typing_extensions==4.7.1
urllib3==2.6.3
Werkzeug==3.1.6
WTForms==3.2.1
